+++ src/components/Calendar/CustomEvents/CustomEventDialog.js
@@ -15,11 +15,11 @@ import {
 import { withStyles } from '@material-ui/core/styles';
 import { Add, Edit } from '@material-ui/icons';
 import PropTypes from 'prop-types';
-import { addCustomEvent, editCustomEvent } from '../../actions/AppStoreActions';
+import { addCustomEvent, editCustomEvent } from '../../../actions/AppStoreActions';
 import ScheduleSelector from './ScheduleSelector';
 import ReactGA from 'react-ga';
-import analyticsEnum, { logAnalytics } from '../../analytics';
-import { isDarkMode } from '../../helpers';
+import analyticsEnum, { logAnalytics } from '../../../analytics';
+import { isDarkMode } from '../../../helpers';
 
 const styles = () => ({
     container: {
